# services/route-service/internal/api/v1/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API serves as an interface to interact with the My Route Service platform, providing endpoints for managing bus routes, Stops, and schedules.",
        "title": "My Route Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Mohammad Mizan",
            "url": "http://swagger.io/support",
            "email": "takbir.jcd@gmail.com"
        },
        "license": {
            "name": "Apache License Version 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/routes/{routeId}/stops": {
            "get": {
                "description": "Retrieve a list of stops for a given route.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stops"
                ],
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Route ID",
                        "name": "routeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "stops",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.StopResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid route ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Route not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new stop to the specified route.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stops"
                ],
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Route ID",
                        "name": "routeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Stop information",
                        "name": "stop",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddStopRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "createdStop",
                        "schema": {
                            "$ref": "#/definitions/dto.StopResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or incorrect data",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Route not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    }
                }
            }
        },
        "/routes/{routeId}/stops/{id}": {
            "get": {
                "description": "Retrieve details of a stop by its unique ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stops"
                ],
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Route ID",
                        "name": "routeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Stop ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "stop",
                        "schema": {
                            "$ref": "#/definitions/dto.StopResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid route ID or stop ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Route or stop not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    }
                }
            },
            "put": {
                "description": "Update details of a stop for a given route.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stops"
                ],
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Route ID",
                        "name": "routeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Stop ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated stop information",
                        "name": "updateStopRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateStopRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updatedStop",
                        "schema": {
                            "$ref": "#/definitions/dto.StopResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid route ID or stop ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Route or stop not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a stop from a given route.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stops"
                ],
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Route ID",
                        "name": "routeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Stop ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Invalid route ID or stop ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Route or stop not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.AddStopRequest": {
            "type": "object",
            "required": [
                "name",
                "sequence"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "sequence": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ScheduleResponse": {
            "type": "object",
            "properties": {
                "arrival_time": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "departure_time": {
                    "type": "string"
                },
                "schedule_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.StopResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ScheduleResponse"
                    }
                },
                "sequence": {
                    "type": "integer"
                },
                "stop_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateStopRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "sequence": {
                    "type": "integer"
                }
            }
        },
        "pkg.ErrorMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}